+++ src/migrations/Install.php
@@ -130,12 +130,13 @@ class Install extends Migration
                 'podcastIsNewFeedUrl' => $this->boolean(),
                 'copyright' => $this->string(1000),
                 'medium' => $this->string(50),
+                'locked' => $this->boolean(),
                 'dateCreated' => $this->dateTime()->notNull(),
                 'dateUpdated' => $this->dateTime()->notNull(),
                 'uid' => $this->uid(),
             ]);
 
-            $this->createIndex(null, '{{%studio_i18n}}', ['publishOnRSS', 'medium'], false);
+            $this->createIndex(null, '{{%studio_i18n}}', ['publishOnRSS', 'medium', 'locked'], false);
 
             $this->addForeignKey(null, '{{%studio_i18n}}', ['elementId'], '{{%elements}}', ['id'], 'CASCADE', null);
             $this->addForeignKey(null, '{{%studio_i18n}}', ['siteId'], '{{%sites}}', ['id'], 'CASCADE', 'CASCADE');