+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Added <podcast:medium> support.
+- Added <podcast:locked> support.
 
 ## 0.7.0 - 2023-05-23